Noise more or less works, other than this weird lack of randomness at the start.

+++ include/poly/state.h
@@ -46,6 +46,6 @@ extern uint64_t poly_time;
 
 // used for rand_r
 extern unsigned int poly_seed;
-extern float poly_rand_freq; // Kept separate as it updates slowly
+extern unsigned int poly_rand_freq; // Kept separate as it updates slowly
 extern int poly_rand_cnt;
 #endif
